+++ RevitAddin.DA.Tester/Revit/App.cs
@@ -1,31 +1,31 @@
 ﻿using Autodesk.Revit.ApplicationServices;
 using Autodesk.Revit.DB;
+using Revit.DesignApplication;
 using RevitAddin.DA.Tester.Services;
 using System;
 
 namespace RevitAddin.DA.Tester.Revit
 {
-    public class App : IExternalDBApplication
+    public class App : DesignApplication
     {
-        IDisposable designAutomation;
-        public ExternalDBApplicationResult OnStartup(ControlledApplication application)
+        public override void OnStartup()
         {
             Console.WriteLine("----------------------------------------");
             Console.WriteLine($"FullName: \t{this.GetType().Assembly.FullName}");
             Console.WriteLine("----------------------------------------");
             Console.WriteLine($"Location: {this.GetType().Assembly.Location}");
             Console.WriteLine("----------------------------------------");
-            Console.WriteLine($"AddInName: \t{application.ActiveAddInId?.GetAddInName()}");
+            Console.WriteLine($"AddInName: \t{Application.ActiveAddInId?.GetAddInName()}");
             Console.WriteLine("----------------------------------------");
-
-            designAutomation = new DesignAutomationLoadVersion<DesignAutomationController>();
-            return ExternalDBApplicationResult.Succeeded;
         }
 
-        public ExternalDBApplicationResult OnShutdown(ControlledApplication application)
+        public override void OnShutdown()
         {
-            designAutomation?.Dispose();
-            return ExternalDBApplicationResult.Succeeded;
+            
+        }
+        public override bool Execute(Application application, string filePath, Document document)
+        {
+            return new DesignAutomationController().Execute(application, filePath, document);
         }
     }
 }