+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] / 2025-09-18
+### Features
+- Create `Revit.DesignApplication` with `DesignApplication` to load correct version of Revit.
+### Updates
+- Use `DesignApplication` to load correct version of Revit.
+
 ## [1.4.0] / 2025-08-27
 ### Features
 - Support Bundle multiple versions of Revit using `DesignAutomationLoadVersion`. (#7)
@@ -40,6 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.4.1]: ../../compare/1.4.0...1.4.1
 [1.4.0]: ../../compare/1.3.1...1.4.0
 [1.3.1]: ../../compare/1.3.0...1.3.1
 [1.3.0]: ../../compare/1.2.0...1.3.0